fe/requirements_venv.txt
python-dotenv==1.0.0
streamlit==1.24.1
